\docType{methods}
\name{calibration-matrix}
\alias{calibration-matrix}
\alias{calibrations}
\alias{print.cal}
\alias{print.cal.pbo13}
\alias{print.cal.pbo13.AFC34}
\alias{print.cal.pbo13.ANC33}
\alias{print.cal.pbo13.ANC34}
\alias{print.cal.pbo13.err}
\alias{print.cal.pbou}
\alias{print.cal.roel1}
\alias{print.cal.roel2}
\alias{print.cal.surf}
\alias{print.cal.surf.AS33}
\alias{print.cal.surf.AS43}
\title{Calibration matrices for BSM gauge data}
\usage{
  \method{print}{cal} (x, ...)

  \method{print.cal}{roel1} (x, ...)

  \method{print.cal}{roel2} (x, ...)

  \method{print.cal}{pbou} (x, ...)

  \method{print.cal}{pbo13} (x, ...)

  \method{print.cal.pbo13}{err} (x, ...)

  \method{print.cal.pbo13}{ANC34} (x, ...)

  \method{print.cal.pbo13}{ANC33} (x, ...)

  \method{print.cal.pbo13}{AFC34} (x, ...)

  \method{print.cal}{surf} (x, ...)

  \method{print.cal.surf}{AS33} (x, ...)

  \method{print.cal.surf}{AS43} (x, ...)
}
\arguments{
  \item{x}{object to print}

  \item{...}{additional parameters}
}
\description{
  Calibration matrices for BSM gauge data
}
\author{
  A.J. Barbour <andy.barbour@gmail.com>
}

