% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gain}
\alias{gain}
\alias{gain.default}
\title{Gain correction}
\usage{
gain(X, gap = NULL, ref.strn = FALSE, ...)

\method{gain}{default}(X, gap = NULL, ref.strn = FALSE, ...)
}
\arguments{
\item{X}{numeric; gauge strains}

\item{gap}{numeric or character; the gap distance}

\item{ref.strn}{logical; should the results be centered?}

\item{...}{additional parameters}

\item{constant}{numeric; value to use for de-linearizing}

\item{unref.strn}{logical; should the centering be reversed?}
}
\description{
Gain correction
}
