% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gap_distance}
\alias{gap_distance}
\title{Choose a GTSM gap distance}
\usage{
gap_distance(gap = NULL)
}
\arguments{
\item{gap}{character; the type of gap}
}
\description{
Choose a GTSM gap distance
}
\details{
The gap is either \code{"one"} representing 100 microns, or
\code{"two"} for 200 microns.
}

