% Generated by roxygen2 (4.0.2): do not edit by hand
\name{max_shear}
\alias{max_shear}
\alias{max_shear.default}
\title{Calculate maximum horizontal shear from horizontal tensor strains}
\usage{
max_shear(E11, E22, E12, ...)

\method{max_shear}{default}(E11, E22, E12, ...)
}
\arguments{
\item{E11}{numeric; 1-axis strain}

\item{E22}{numeric; 2-axis strain}

\item{E12}{numeric; Shear strains}

\item{...}{additional parameters}
}
\description{
Calculate maximum horizontal shear from horizontal tensor strains
}

