\docType{data}
\name{minphs}
\alias{minphs}
\title{Minimum phase lowpass filter weights for strain data}
\format{A list with filter weights by decimation factor}
\description{
  Minimum phase lowpass filter weights for strain data
}
\references{
  D.C. Agnew and K. Hodgkinson (2007), Designing compact
  causal digital filters for low-frequency strainmeter
  data, Bulletin of the Seismological Society of America,
  vol 97, 1B, 1-99, doi: 10.1785/0120060088
}
\keyword{datasets}

