% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R
\name{orientation_conventions}
\alias{orientation_conventions}
\title{Remind, with ASCII art, the gauge numbering/naming conventions}
\usage{
orientation_conventions()
}
\description{
Remind, with ASCII art, the gauge numbering/naming conventions
}

