\docType{data}
\name{strain-constants}
\alias{constants}
\alias{strain-constants}
\title{Constants used as defaults}
\format{List of 3
 $ Poisson    :List of 3
 $ conversions:List of 2
 $ bsm        :List of 5}
\usage{
  .constants

  constants(do.str = TRUE)
}
\arguments{
  \item{do.str}{logical; should the structure be printed?}
}
\description{
  Constants used as defaults
}
\details{
  The helper function \code{\link{constants}} shows (the
  structure of, optionally) and returns \code{.constants}.

  The following constants are assumed \describe{
  \item{Poisson's ratios}{ (Poisson solid)} }
}
\seealso{
  \code{\link{strain}}
}
\keyword{datasets}

