\docType{methods}
\name{strain-getters}
\alias{get_az}
\alias{get_itype}
\alias{strain-getters}
\title{Functions to get information about a \code{bsm} object.}
\usage{
  get_az(B)

  get_itype(B)
}
\arguments{
  \item{B}{object with class\code{bsm}}
}
\description{
  Functions to get information about a \code{bsm} object.
}
\author{
  A.J. Barbour <andy.barbour@gmail.com>
}

