\docType{methods}
\name{strain-miscellany}
\alias{circle}
\alias{strain-miscellany}
\title{Miscellaneous utilities}
\usage{
circle(dtheta = 1, origin = 0)
}
\arguments{
  \item{dtheta}{numeric; the disretization in degrees.}

  \item{origin}{numeric; where the first point begins, in
  degrees}
}
\description{
Miscellaneous utilities

Return a unit circle.
}
\details{
Origin is for a right-hand (ccw) system.
}
\author{
A.J. Barbour <andy.barbour@gmail.com>
}

