\docType{package}
\name{strain-package}
\alias{strain-package}
\alias{strain,}
\alias{strainmeter}
\title{Tools to work with geodetic strainmeter data.}
\description{
  Some description
}
\details{
  Some details
}
\author{
  Andrew J. Barbour <andy.barbour@gmail.com>
}
\seealso{
  \code{\link{bsm-methods}}
}

