% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresses.R
\name{stresses}
\alias{stresses}
\alias{principal_stresses}
\title{Compute plane stress tensor from plane strain tensor}
\usage{
stresses(E11, E22, E12, G = 3e+10, nu = 0.25,
  extension.positive = TRUE, as.mat = TRUE)

principal_stresses(S)
}
\arguments{
\item{E11, E22, E12}{numeric; strain components in 1-2 coordinate system}

\item{G}{numeric; elastic shear modulus; 30 GPa is a reasonable value
for deep crustal rock}

\item{nu}{numeric; drained Poissons ratio; 1/4 is a reasonable value
for a Poissonian solid}

\item{extension.positive}{logical; positive extensions are extensional}

\item{as.mat}{logical; should the result be returned as a matrix?}

\item{S}{symmetric stress tensor}
}
\description{
Compute plane stress tensor from plane strain tensor
}
\seealso{
\code{\link{CoulombStress}}
}
