% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R
\name{as.bsm}
\alias{as.bsm}
\alias{as.bsm.default}
\alias{bsm}
\alias{is.bsm}
\alias{is.raw_strain}
\title{Convert an object to one with class \code{'bsm'}}
\usage{
as.bsm(X, ...)

\method{as.bsm}{default}(X, g0az = 0, station = "", i.type = "GTSM",
  linearize = TRUE, cal.tbl = c("pbou"), ...)

is.bsm(X)

is.raw_strain(X)
}
\arguments{
\item{X}{generic object}

\item{...}{additional arguments}

\item{g0az}{numeric; the orientation of the zeroth gauge}

\item{station}{character; name of the station associated with the data}

\item{i.type}{character; instrument type}

\item{linearize}{logical; Should \code{X} be linearized?}

\item{cal.tbl}{character; retrieve calibration coefficients from this table}
}
\description{
Convert an object to one with class \code{'bsm'}
}

