\name{as.bsm}
\alias{as.bsm}
\alias{bsm}
\alias{is.bsm}
\alias{print.bsm}
\title{Borehole strainmeter (bsm) methods and classes}
\usage{
  as.bsm(X, i.type = "gtsm", g0az = 0, ...)

  print.bsm(B, ...)

  is.bsm(X)
}
\arguments{
  \item{X}{generic object}

  \item{i.type}{character; instrument type}

  \item{B}{object of class \code{bsm}}

  \item{...}{additional arguments}
}
\description{
  Borehole strainmeter (bsm) methods and classes

  Logical test for 'bsm' class
}

