% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calibMatrix}
\alias{calibMatrix}
\alias{calibMatrix.default}
\title{Return a calibration matrix, inverted or not}
\usage{
calibMatrix(MC, ...)

\method{calibMatrix}{default}(MC, MD = NULL, MH = NULL, axis_deg = 0,
  delta_deg = 0, invert = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{MC}{numeric;}

\item{MD}{numeric;}

\item{MH}{numeric;}

\item{axis_deg}{numeric;}

\item{delta_deg}{numeric;}

\item{pinvert}{logical; should the matrix be (pseudo-)inverted?}

\item{...}{additional parameters}
}
\description{
Return a calibration matrix, inverted or not
}
\details{
The argument \code{MC} may be a matrix 4with three columns representing
\code{MC}, \code{MD}, and \code{MH} respectively.  If that's the case, any
arguments given for \code{MD}, and \code{MH} are ignored.

Coefficients are in the form presented by Roeloffs 2010.

The pseudo-inverse is computed using \code{corpcor::pseudoinverse} with
default tolerances.
}

