% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\alias{calibrate.default}
\title{Apply a calibration to strain data}
\usage{
calibrate(B, Cmat, invert = FALSE)

\method{calibrate}{default}(B, Cmat, invert = FALSE)
}
\arguments{
\item{B}{calibratee}

\item{Cmat}{calibrator}

\item{invert}{logical; should \code{Cmat} be (pseudo) inverted?}

\item{...}{additional objects}
}
\description{
Apply a calibration to strain data
}
