\docType{methods}
\name{calibration-matrix}
\alias{calibration-matrix}
\alias{calibrations}
\alias{print.cal}
\alias{print.cal.pbo13}
\alias{print.cal.pbo13.AFC34}
\alias{print.cal.pbo13.ANC33}
\alias{print.cal.pbo13.ANC34}
\alias{print.cal.pbo13.err}
\alias{print.cal.pbou}
\alias{print.cal.roel1}
\alias{print.cal.roel2}
\alias{print.cal.surf}
\alias{print.cal.surf.AS33}
\alias{print.cal.surf.AS43}
\title{Calibration matrices for BSM gauge data}
\usage{
  print.cal(x, ...)

  print.cal.roel1(x, ...)

  print.cal.roel2(x, ...)

  print.cal.pbou(x, ...)

  print.cal.pbo13(x, ...)

  print.cal.pbo13.err(x, ...)

  print.cal.pbo13.ANC34(x, ...)

  print.cal.pbo13.ANC33(x, ...)

  print.cal.pbo13.AFC34(x, ...)

  print.cal.surf(x, ...)

  print.cal.surf.AS33(x, ...)

  print.cal.surf.AS43(x, ...)
}
\arguments{
  \item{x}{object to print}

  \item{...}{additional parameters}
}
\description{
  Calibration matrices for BSM gauge data
}
\author{
  A.J. Barbour <andy.barbour@gmail.com>
}

