\docType{methods}
\name{calibration-matrix}
\alias{calibration-matrix}
\alias{calibrations}
\alias{print.cal}
\alias{print.cal.pbo}
\alias{print.cal.pbou}
\alias{print.cal.roel1}
\alias{print.cal.roel2}
\alias{print.cal.surf1}
\alias{print.cal.surf2}
\title{Calibration matrices for BSM gauge data}
\usage{
  print.cal(X, ...)

  print.cal.roel1(X, ...)

  print.cal.roel2(X, ...)

  print.cal.pbou(X, ...)

  print.cal.pbo(X, ...)

  print.cal.surf1(X, ...)

  print.cal.surf2(X, ...)
}
\description{
  Calibration matrices for BSM gauge data
}
\author{
  A.J. Barbour <andy.barbour@gmail.com>
}

