% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{geod_rotate}
\alias{geod_rotate}
\alias{geod_rotate.default}
\alias{geod_rotate2}
\title{Transform tensor strains in a 1-2 system into an r-t system}
\usage{
geod_rotate(E11, E22, E12, geodesic.deg = 0, NsEwNw = FALSE, ...)

geod_rotate2(BSM, geodesic.deg = 0, check.columns = FALSE, ...)

\method{geod_rotate}{default}(E11, E22, E12, geodesic.deg = 0,
  NsEwNw = FALSE, ...)
}
\arguments{
\item{E11}{numeric; 1-axis strain}

\item{E22}{numeric; 2-axis strain}

\item{E12}{numeric; shear strain}

\item{geodesic.deg}{numeric; the azimuth of the geodesic (in degrees)}

\item{NsEwNw}{unused??}

\item{...}{additional parameters}

\item{BSM}{numeric matrix with columns \code{c("ar","gam1","gam2")}}

\item{check.columns}{logical; should column names be checked?}
}
\description{
Transform tensor strains in a 1-2 system into an r-t system
}
\details{
\code{\link{geod_rotate2}} is a kludge for the moment, to
prove tensor strain rotations
}

