% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.calibrated}
\alias{is.calibrated}
\alias{is.calibrated.default}
\title{Test whether strain is calibrated}
\usage{
is.calibrated(X)

\method{is.calibrated}{default}(X)
}
\arguments{
\item{X}{object to test}
}
\description{
Test whether strain is calibrated
}

