% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{lpfilter}
\alias{lpfilter}
\alias{lpdfilter}
\alias{lpfilter.dec}
\title{Causal lowpass filtering with decimation}
\usage{
lpfilter(y, wgts, ndec = 1)

lpfilter.dec(y, frq = 1, ndecs = c(2, 5), plt = FALSE, ...)
}
\arguments{
\item{y}{numeric}

\item{wgts}{numeric; convolution filter weights}

\item{ndec}{integer; the decimation factors to use}

\item{frq}{integer; the sampling frequency of \code{y}}

\item{plt}{logical; should the decimation stages be plotted?}

\item{...}{additional parameters to \code{\link{ts}} in the case \code{y}
does not have class \code{'ts'}.}
}
\description{
Causal lowpass filtering with decimation
}
\details{
Applies a single-sided convolution filter with the filter
weights (\code{wgts}) to the data (\code{y}).  The decimation factor
is set with \code{ndec} but is only applied when \code{ndec>1}.
}
\seealso{
\code{\link{straight_decimate}}, \code{\link{filter}}, \code{\link{strain-filtering}}
}
