% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strains.R
\name{principals}
\alias{principals}
\alias{principals.default}
\alias{principals.lsm3}
\alias{principals.lsm2}
\title{Calculate principal strains}
\usage{
principals(S, ...)

\method{principals}{default}(S, B, alpha, is.radians = TRUE, ...)

\method{principals}{lsm3}(S, ...)

\method{principals}{lsm2}(S, ...)
}
\arguments{
\item{S}{strain information}

\item{...}{additional parameters}
}
\description{
Strainmeters are generally designed to resemble common
strain-gauge rosette configurations to simplify gauge
combinations.
}
\details{
If there are only two components of the LSM the angle of maximum shear
cannot be solved for.

The calculation is based on 2-D Mohr's circle  with the 1-axis being normal
and the 2-axis being shear.
}
\section{Rosette classifications}{

\itemize{
  \item{BSM}{T-Delta}
  \item{LSM-3}{Rectangular three-gauge \eqn{45^{\circ}}, star shaped}
  \item{LSM-2}{Rectangular two-gauge \eqn{90^{\circ}}}
}
}

\references{
Meir, J.H. (19XX),"Strain Rosettes" , in XX, pp. 412, edited by
Altamani 19XX
}
