% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain-package.R
\docType{data}
\name{strain-constants}
\alias{.strainconstants}
\alias{strain-constants}
\alias{strain.constants}
\title{Constants used as defaults}
\format{An object of class \code{list} of length 3.}
\usage{
.strainconstants

strain.constants(do.str = TRUE)
}
\arguments{
\item{do.str}{logical; should the structure be printed?}
}
\description{
Constants used as defaults
}
\details{
The helper function \code{\link{strain.constants}}
shows (the structure of, optionally)
and returns \code{.strainconstants}.

The following constants are assumed
\describe{
\item{Poisson's ratios}{ (Poisson solid)}
}
}
\seealso{
\code{\link{strain}}
}
\keyword{datasets}

