% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\docType{methods}
\name{strain-getters}
\alias{strain-getters}
\alias{get_az}
\alias{get_station}
\alias{get_itype}
\title{Functions to get information about a \code{bsm} object.}
\usage{
get_az(B)

get_station(B)

get_itype(B)
}
\arguments{
\item{B}{object with class\code{bsm}}

\item{strn.type}{character; the type of strain to retrieve from \code{B}}

\item{...}{additional arguments}
}
\author{
A.J. Barbour <andy.barbour@gmail.com>
}
