\docType{methods}
\name{strain-getters}
\alias{get_az}
\alias{get_itype}
\alias{get_station}
\alias{get_strains}
\alias{strain-getters}
\title{Functions to get information about a \code{bsm} object.}
\usage{
  get_strains(B, ...)

  get_az(B)

  get_station(B)

  get_itype(B)
}
\arguments{
  \item{B}{object with class\code{bsm}}

  \item{strn.type}{character; the type of strain to
  retrieve from \code{B}}

  \item{...}{additional arguments}
}
\description{
  Functions to get information about a \code{bsm} object.
}
\author{
  A.J. Barbour <andy.barbour@gmail.com>
}

