\docType{package}
\name{strain-package}
\alias{strain,}
\alias{strain-package}
\alias{strainmeter}
\title{Tools to work with geodetic strainmeter data.}
\description{
  Some description
}
\details{
  Some details
}
\author{
  Andrew J. Barbour <andy.barbour@gmail.com>
}
\references{
  Grant, E. B. (2010), Gladwin Tensor Strainmeter
  Calibration Using Seismic Data: Instrument Calibration
  Methods and Wave Gradiometry Applications, PhD thesis,
  University of Memphis.

  Hart, R. H. G., M. T. Gladwin, R. L. Gwyther, D. C.
  Agnew, and F. K. Wyatt (1996), Tidal calibration of
  borehole strain meters: Removing the effects of
  small-scale inhomogeneity, J. Geophys. Res., 101(B11),
  25553-25571, doi:10.1029/96JB02273.

  Hodgkinson, K., J. Langbein, B. Henderson, D. Mencin, and
  A. Borsa (2013), Tidal calibration of plate boundary
  observatory borehole strainmeters, J. Geophys. Res. Solid
  Earth, 118, 447-458, doi:10.1029/2012JB009651.

  Langbein, J. (2010), Effect of error in theoretical Earth
  tide on calibration of borehole strainmeters, Geophys.
  Res. Lett., 37, L21303, doi:10.1029/2010GL044454.

  Roeloffs, E. (2010), Tidal calibration of Plate Boundary
  Observatory borehole strainmeters: Roles of vertical and
  shear coupling, J. Geophys. Res., 115, B06405,
  doi:10.1029/2009JB006407.
}
\seealso{
  \code{\link{bsm-methods}}
}

