% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strains}
\alias{strains}
\title{Extract strains from an abject}
\usage{
strains(B, ...)
}
\arguments{
\item{B}{object}

\item{...}{additional parameters}
}
\description{
Extract strains from an abject
}
\seealso{
\code{\link{bsm-methods}}
}

